pyproject.toml
[project]
name = "cos-rewrite-mlflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.11"
dependencies = [
    "graphdatascience>=1.18",
    "lets-plot>=4.8.2",
    "lifelines>=0.30.0",
    "mlflow>=3.8.1",
    "neo4j>=6.0.3",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pymer4>=0.9.2",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "statsmodels>=0.14.6",
]
